requirements.txt
pyttsx3>=2.90
faster-whisper>=0.9.0
google-generativeai>=0.3.1
sqlite3-api>=1.0.1
tk>=0.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
sounddevice>=0.4.6
soundfile>=0.12.1
TTS>=0.17.0
chromadb>=0.4.18
huggingface_hub>=0.19.0
langchain>=0.0.267
sentence-transformers>=2.2.2
llama-cpp-python>=0.2.0
pillow>=10.0.0 